requirements.txt
python-binance==1.0.19
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas>=2.2.0
requests==2.31.0
python-dotenv==1.0.0 